Using dependencies environment variable

+++ project_tests.sh
@@ -1,7 +1,13 @@
 #!/usr/bin/env bash
 set -e
 
-composer install
+: "${dependencies:?Need to set dependencies environment variable}"
+if [ "$dependencies" = "lowest" ]; then
+    composer update --prefer-lowest --no-interaction
+    proofreader src/ test/
+else
+    composer update --no-interaction
+fi
 rm -f build/*.xml
 proofreader src/ tests/
 vendor/bin/phpunit --log-junit build/phpunit.xml
